Added casting to float for prices

+++ app/Repositories/CompanyRepository.php
@@ -115,11 +115,11 @@ class CompanyRepository extends BaseRepository
     private function formatPricesForDb($input)
     {
         if(isset($input['delivery_costs'])) {
-            $input['delivery_costs'] = str_replace(',', '.', $input['delivery_costs']);
+            $input['delivery_costs'] = (float)str_replace(',', '.', $input['delivery_costs']);
         }
 
         if(isset($input['min_order_amount'])) {
-            $input['min_order_amount'] = str_replace(',', '.', $input['min_order_amount']);
+            $input['min_order_amount'] = (float)str_replace(',', '.', $input['min_order_amount']);
         }
         return $input;
     }
